LF-3913 - Complete logic fix and cleanup, rename some exports
