+++ packages/webapp/src/App.jsx
@@ -14,7 +14,7 @@
  */
 
 import { Suspense } from 'react';
-import NavBar from './containers/Navigation';
+import Navigation from './containers/Navigation';
 import history from './history';
 import Routes from './Routes.jsx';
 import clsx from 'clsx';
@@ -32,7 +32,7 @@ function App() {
       </Suspense>
       <div className={clsx(styles.mainColumn)}>
         <Suspense fallback={null}>
-          <NavBar history={history} />
+          <Navigation history={history} />
         </Suspense>
         <div className={styles.app}>
           <OfflineDetector />
