--- packages/webapp/src/components/Navigation/NavBar/styles.module.scss
@@ -1,90 +1,0 @@
-/*
- *  Copyright 2023 LiteFarm.org
- *  This file is part of LiteFarm.
- *
- *  LiteFarm is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation, either version 3 of the License, or
- *  (at your option) any later version.
- *
- *  LiteFarm is distributed in the hope that it will be useful;
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- *  GNU General Public License for more details, see <https://www.gnu.org/licenses/>.
- */
-
-.menuButton {
-  position: absolute;
-  padding: 0;
-  left: 24px;
-
-  &:hover {
-    background-color: transparent;
-  }
-
-  @media only screen and (min-width: 415px) {
-    display: none;
-  }
-}
-
-.icons {
-  display: flex;
-  position: absolute;
-  right: 8px;
-}
-
-.appBar {
-  background-color: var(--bgInputListTile);
-  box-shadow: 0px 2px 0px 0px #0000000D;
-  height: var(--global-navbar-height);
-}
-
-.toolbar {
-  display: flex;
-  justify-content: flex-start;
-  flex-grow: 1;
-  align-items: center;
-}
-
-.centerContent {
-  justify-content: center;
-}
-
-.notificationButton {
-  transform: translateY(1px);
-}
-
-.iconButton {
-  margin: 8px;
-  padding: 0;
-}
-
-.burgerMenu {
-  font-size: 32px;
-  color: var(--teal700);
-
-  @media only screen and (min-width: 900px) {
-    font-size: 40px;
-  }
-}
-
-.green {
-  color: var(--teal700);
-}
-
-.p {
-  margin-bottom: 12px;
-}
-
-.black {
-  color: var(--grey900);
-}
-
-.logo {
-  cursor: pointer;
-  margin-left: var(--global-sidebar-width);
-}
-
-.displayNone {
-  display: none;
-}
