--- packages/webapp/src/components/Navigation/NoFarmNavBar/styles.module.scss
@@ -1,50 +1,0 @@
-/*
- *  Copyright (C) 2007 Free Software Foundation, Inc. <https://fsf.org/>
- *  This file (styles.scss) is part of LiteFarm.
- *
- *  LiteFarm is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation, either version 3 of the License, or
- *  (at your option) any later version.
- *
- *  LiteFarm is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- *  GNU General Public License for more details, see <https://www.gnu.org/licenses/>.
- */
-@import '../../../assets/mixin';
-
-.navBar {
-  top: 0;
-  background: white;
-  width: 100%;
-  max-width: 1024px;
-  height: 76px;
-  position: relative;
-  & .itemContainer {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    -ms-transform: translate(-50%, -50%);
-    transform: translate(-50%, -50%);
-  }
-}
-
-.middleLogo {
-  height: 62px;
-}
-
-@media (max-width: 800px) {
-  .actionItemContainer {
-    margin-top: 16px;
-    margin-right: 0;
-  }
-
-  .navBar {
-    min-height: 76px;
-  }
-
-  .middleLogo {
-    height: 58px;
-  }
-}
